dip/requirement.txt
tensorflow
opencv-python
scikit-image
pillow
pixellib